requirements-dev.txt
-r requirements.txt

dogpile.cache==0.5.6

# Test
coveralls==0.5
flake8==2.4.1
py==1.4.30  # Explicit for Travis work-around
pytest-cache==1.0
pytest-cov==1.8.1
pytest==2.7.2
python-dotenv==0.1.3
tox==2.1.1
